Add component info item to model "More" menu
Since we deleted the "Components" panel, we need a way for a user to
get information about a component before instantiating it. This menu
selection displays a list of components. When the user can selects a
component from the list, a ComponentInfoDialogBox is displayed, showing
basic information and a link to the component page on the CSDMS website.