Add new VisualizeFile class and run/visualize/uuid API call
VisualizeFile takes a uuid, from which the model driver component is
obtained. If the driver has a parameter `_visualization`, the file it
references is displayed in the browser when the *run/visualize/uuid*
API call is made. If this parameter isn't present, a directory listing
of the component output is displayed.

One problem: if an invalid path is entered for `_visualization`, a 404
Not Found page is displayed. I should handle this.