Update stage_component to use model_metadata package
The model_metadata package is used by the new PyMT. It uses jinja
templates for the component configuration files, which is
incompatible with the old home-grown templating used in WMT.

Here, I've modified the stage_component function to use the
FileSystemLoader class from model_metadata to fill in the templated
values of a component with values taken from the WMT server (and the
WMT client). I've removed the old _component_stagein function.